Move responsibility for stream setup into client
